mozhisense-backend/requirements.txt
numpy<2.0.0
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.2
stanza==1.8.2
indic-nlp-library==0.92
inltk==0.9
torch==2.2.2
beautifulsoup4==4.12.3
